# api-gateway/api/docs/swagger.yaml
definitions:
  models.Comments:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      owner_id:
        type: string
      post_id:
        type: string
      text:
        type: string
      updated_at:
        type: string
    type: object
  models.Error:
    properties:
      message:
        type: string
    type: object
  models.Post:
    properties:
      categories:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comments'
        type: array
      content:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      dislikes:
        type: integer
      id:
        type: string
      likes:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      views:
        type: integer
    type: object
  models.PostComments:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      owner_id:
        type: string
      post_id:
        type: string
      text:
        type: string
      updated_at:
        type: string
    type: object
...
  models.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a example of USER SERVICE, POST SERVICE and COMMENT SERVICE.
    Author Sardor Keldimurodov
  title: Welcome to API-GATEWAY
  version: "1.0"
paths:
  /v1/comment/:
    post:
      consumes:
      - application/json
      description: Api for creating a new comment
      parameters:
      - description: createComment
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.PostComments'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostComments'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: CreateComment ...
      tags:
      - comment
  /v1/comment/{id}:
    delete:
      consumes:
      - application/json
      description: Api deletes post
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: Delete Comment by id
      tags:
      - comment
    get:
      consumes:
      - application/json
      description: Api for getting comment by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostComments'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: GetComment
      tags:
      - comment
    put:
      consumes:
      - application/json
      description: Api returns updates comment
      parameters:
      - description: UpdateComment
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.PostComments'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: Update Commment
      tags:
      - comment
  /v1/comments/:
    get:
      consumes:
      - application/json
      description: Api returns list of comments from
      parameters:
      - description: Page
        in: path
        name: page
        required: true
        type: integer
      - description: Limit
        in: path
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: All comments from the service
      tags:
      - comment
...
  /v1/post/:
    post:
      consumes:
      - application/json
      description: Api for creating a new post
      parameters:
      - description: createPost
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: CreatePost ...
      tags:
      - post
  /v1/post/{id}:
    delete:
      consumes:
      - application/json
      description: Api deletes post
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: Delete Post
      tags:
      - post
    get:
      consumes:
      - application/json
      description: Api for getting post by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: GetPost
      tags:
      - post
    put:
      consumes:
      - application/json
      description: Api returns updates post
      parameters:
      - description: UpdatePost
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: UpdatePost
      tags:
      - post
  /v1/posts/:
    get:
      consumes:
      - application/json
      description: Api returns list of posts
      parameters:
      - description: Page
        in: path
        name: page
        required: true
        type: integer
      - description: Limit
        in: path
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: All posts
      tags:
      - post
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"